#471 Allow configuring custom environment and component locator in application factory
