+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/core/context/ApplicationEnvironment.java
@@ -17,42 +17,22 @@
 
 package org.dockbox.hartshorn.core.context;
 
-import org.dockbox.hartshorn.core.HartshornUtils;
 import org.dockbox.hartshorn.core.boot.beta.HartshornApplicationManager;
 import org.dockbox.hartshorn.core.context.element.TypeContext;
 
 import java.lang.annotation.Annotation;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 
-import lombok.Getter;
+public interface ApplicationEnvironment {
 
-public class ApplicationEnvironment {
+    PrefixContext prefixContext();
 
-    @Getter private final PrefixContext prefixContext;
-    @Getter private final boolean isCI;
-    @Getter private final HartshornApplicationManager application;
+    boolean isCI();
 
-    public ApplicationEnvironment(final Collection<String> prefixes, final HartshornApplicationManager application) {
-        this.application = application;
-        this.isCI = this.detectCI();
-        this.prefixContext = new PrefixContext(prefixes, this);
-        this.application().log().debug("Created new application environment (isCI: %s, prefixCount: %d)".formatted(this.isCI(), this.prefixContext().prefixes().size()));
-    public HartshornApplicationManager manager() {
-        // TODO: Temporary placeholder
-        return this.application();
-    }
+    public HartshornApplicationManager manager();
 
-    private boolean detectCI() {
-        return HartshornUtils.isCI();
-    }
-
-    public void prefix(final String prefix) {
-        this.prefixContext.prefix(prefix);
-    }
-    }
-
+    void prefix(String prefix);
 
     /**
      * Gets types decorated with a given annotation, both classes and annotationsWith. The prefix is
@@ -66,13 +46,9 @@ public class ApplicationEnvironment {
      *
      * @return The annotated types
      */
-    public <A extends Annotation> Collection<TypeContext<?>> types(final Class<A> annotation) {
-        return this.prefixContext.types(annotation);
-    }
+    <A extends Annotation> Collection<TypeContext<?>> types(final Class<A> annotation);
 
-    public <A extends Annotation> Collection<TypeContext<?>> types(final String prefix, final Class<A> annotation, final boolean skipParents) {
-        return this.prefixContext.types(prefix, annotation, skipParents);
-    }
+    <A extends Annotation> Collection<TypeContext<?>> types(final String prefix, final Class<A> annotation, final boolean skipParents);
 
     /**
      * Gets types decorated with a given annotation, both classes and annotationsWith. The prefix is
@@ -88,9 +64,7 @@ public class ApplicationEnvironment {
      *
      * @return The annotated types
      */
-    public <A extends Annotation> Collection<TypeContext<?>> types(final Class<A> annotation, final boolean skipParents) {
-        return this.prefixContext.types(annotation, skipParents);
-    }
+    <A extends Annotation> Collection<TypeContext<?>> types(final Class<A> annotation, final boolean skipParents);
 
     /**
      * Gets all sub-types of a given type. The prefix is typically a package. If no sub-types exist
@@ -103,13 +77,9 @@ public class ApplicationEnvironment {
      *
      * @return The list of sub-types, or an empty list
      */
-    public <T> Collection<TypeContext<? extends T>> children(final TypeContext<T> parent) {
-        return this.prefixContext.children(parent);
-    }
+    <T> Collection<TypeContext<? extends T>> children(final TypeContext<T> parent);
 
-    public <T> Collection<TypeContext<? extends T>> children(final Class<T> parent) {
-        return this.prefixContext.children(parent);
-    }
+    <T> Collection<TypeContext<? extends T>> children(final Class<T> parent);
 
     /**
      * Gets supertypes.
@@ -119,35 +89,7 @@ public class ApplicationEnvironment {
      *
      * @return the supertypes
      */
-    public Collection<Class<?>> parents(final Class<?> current) {
-        final Set<Class<?>> supertypes = HartshornUtils.emptySet();
-        final Set<Class<?>> next = HartshornUtils.emptySet();
-        final Class<?> superclass = current.getSuperclass();
+    Collection<Class<?>> parents(final Class<?> current);
 
-        if (Object.class != superclass && null != superclass) {
-            supertypes.add(superclass);
-            next.add(superclass);
-        }
-
-        for (final Class<?> interfaceClass : current.getInterfaces()) {
-            supertypes.add(interfaceClass);
-            next.add(interfaceClass);
-        }
-
-        for (final Class<?> cls : next) {
-            supertypes.addAll(this.parents(cls));
-        }
-
-        return supertypes;
-    }
-
-    public List<Annotation> annotationsWith(final TypeContext<?> type, final Class<? extends Annotation> annotation) {
-        final List<Annotation> annotations = HartshornUtils.emptyList();
-        for (final Annotation typeAnnotation : type.annotations()) {
-            if (TypeContext.of(typeAnnotation.annotationType()).annotation(annotation).present()) {
-                annotations.add(typeAnnotation);
-            }
-        }
-        return annotations;
-    }
+    List<Annotation> annotationsWith(final TypeContext<?> type, final Class<? extends Annotation> annotation);
 }
